# fix: Optimize parser